Online Shopping
The Online Shopping is a web based application intended for online retailers. The main
objective of this application is to make it interactive and its ease of use. It would make
searching, viewing and selection of a product easier. It contains a sophisticated search
engine for user's to search for products specific to their needs. The search engine provides
an easy and convenient way to search for products where a user can Search for a product
interactively and the search engine would refine the products available based on the
user’s input. The user can then view the complete specification of each product. They can
also view the product reviews and also write their own reviews. The application also
provides a drag and drop feature so that a user can add a product to the shopping cart by
dragging the item in to the shopping cart. The main emphasis lies in providing a userfriendly search engine for effectively showing the desired results and its drag and drop
behavior.